+++ src/main/java/de/codescape/jira/plugins/scrumpoker/service/EstimateFieldServiceImpl.java
@@ -16,6 +16,10 @@ import org.springframework.stereotype.Component;
 @Component
 public class EstimateFieldServiceImpl implements EstimateFieldService {
 
+    static final String CUSTOM_FIELD_TYPE_NUMBER = "com.atlassian.jira.plugin.system.customfieldtypes:float";
+    static final String CUSTOM_FIELD_TYPE_TEXT = "com.atlassian.jira.plugin.system.customfieldtypes:textfield";
+    static final String CUSTOM_FIELD_TYPE_TEXTAREA = "com.atlassian.jira.plugin.system.customfieldtypes:textarea";
+
     private final GlobalSettingsService globalSettingsService;
     private final CustomFieldManager customFieldManager;
     private final JiraAuthenticationContext jiraAuthenticationContext;
@@ -42,18 +46,35 @@ public class EstimateFieldServiceImpl implements EstimateFieldService {
     public boolean save(String issueKey, String estimate) {
         ApplicationUser applicationUser = jiraAuthenticationContext.getLoggedInUser();
         MutableIssue issue = issueManager.getIssueByCurrentKey(issueKey);
+        // check permission to save the estimate by the current user
         if (globalSettingsService.load().isCheckPermissionToSaveEstimate() &&
             !permissionManager.hasPermission(ProjectPermissions.EDIT_ISSUES, issue, applicationUser)) {
             errorLogService.logError("User " + applicationUser.getUsername() +
                 " is missing permissions to save estimation for issue " + issueKey + ".", null);
             return false;
         }
-        issue.setCustomFieldValue(findEstimateField(), estimate);
+        // based on the type of the custom field apply the estimate in the correct data type
+        CustomField estimateField = findEstimateField();
+        String estimateFieldKey = estimateField.getCustomFieldType().getKey();
+        switch (estimateFieldKey) {
+            case CUSTOM_FIELD_TYPE_NUMBER:
+                issue.setCustomFieldValue(estimateField, Double.valueOf(estimate));
+                break;
+            case CUSTOM_FIELD_TYPE_TEXT:
+            case CUSTOM_FIELD_TYPE_TEXTAREA:
+                issue.setCustomFieldValue(estimateField, estimate);
+                break;
+            default:
+                errorLogService.logError("Unable to save estimate because field type "
+                    + estimateFieldKey + " is not supported.", null);
+                return false;
+        }
+        // finally update the issue
         try {
             issueManager.updateIssue(applicationUser, issue, UpdateIssueRequest.builder().build());
             return true;
         } catch (RuntimeException e) {
-            errorLogService.logError("Unable to save estimation " + estimate + " for issue "
+            errorLogService.logError("Unable to save estimate " + estimate + " for issue "
                 + issueKey + ". See stacktrace for further details.", e);
             return false;
         }